requirements.txt
django
djangorestframework
celery
redis
python-telegram-bot
django-environ
psycopg[binary]
reportlab
openpyxl
